octofit-tracker/backend/requirements.txt
Django>=4.1
djangorestframework>=3.14
django-cors-headers>=3.13
django-allauth>=0.54
dj-rest-auth>=2.2
djongo>=1.3.6
asgiref>=3.6
sqlparse==0.2.4